+++ packages/dialect-react-ui/components/Chat/screens/Main/ThreadsList/MessagePreview.tsx
@@ -2,7 +2,7 @@ import { useApi, DialectAccount, formatTimestamp } from '@dialectlabs/react';
 import { display } from '@dialectlabs/web3';
 import Avatar from '../../../../Avatar';
 import clsx from 'clsx';
-import { CardinalDisplayAddress } from '../../../../CardinalAddress';
+import { DisplayAddress } from '../../../../DisplayAddress';
 
 type PropsType = {
   dialect: DialectAccount;
@@ -60,15 +60,12 @@ export default function MessagePreview({
       </div>
       <div className="dt-flex dt-grow dt-border-b dt-border-neutral-600 dt-justify-between dt-truncate dt-pr-2">
         <div className="dt-flex dt-flex-col dt-max-w-full dt-truncate">
-          {dialect?.dialect.members.length > 0 &&
-            (program?.provider.connection ? (
-              <CardinalDisplayAddress
-                connection={program?.provider.connection}
-                publicKey={otherMembers[0].publicKey}
-              />
-            ) : (
-              <>{display(otherMembers[0].publicKey)}</>
-            ))}
+          {
+            <DisplayAddress
+              connection={program?.provider.connection}
+              dialectMembers={dialect?.dialect.members}
+            />
+          }
           <FirstMessage dialect={dialect} />
         </div>
         <div className="dt-text-xs dt-opacity-30">