+++ packages/dialect-react-ui/components/Chat/screens/Main/pages/CreateThreadPage/CreateThread.tsx
@@ -1,11 +1,10 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import * as anchor from '@project-serum/anchor';
 import {
   getDialectAddressWithOtherMember,
   ParsedErrorData,
   useApi,
   useDialect,
-  Wallets,
 } from '@dialectlabs/react';
 import clsx from 'clsx';
 import IconButton from '../../../../../IconButton';
@@ -13,11 +12,13 @@ import { A, H1, Input, P } from '../../../../../common/preflighted';
 import { useTheme } from '../../../../../common/ThemeProvider';
 import {
   Button,
+  Footer,
   NetworkBadge,
   Toggle,
   useBalance,
   ValueRow,
 } from '../../../../../common';
+import { fetchAddressFromTwitterHandle } from '../../../../../DisplayAddress';
 import { display } from '@dialectlabs/web3';
 import { Lock, NoLock } from '../../../../../Icon';
 
@@ -28,6 +29,29 @@ interface CreateThreadProps {
   onModalClose?: () => void;
 }
 
+function CardinalCTA() {
+  const { textStyles } = useTheme();
+  return (
+    <P
+      className={clsx(
+        textStyles.small,
+        'dt-opacity-60 dt-text-white dt-text dt-mt-1 dt-px-2'
+      )}
+    >
+      {'Link your Twitter handle with '}
+      <A
+        href={'https://twitter.cardinal.so'}
+        target="_blank"
+        rel="noreferrer"
+        className="dt-underline"
+      >
+        twitter.cardinal.so
+      </A>
+      {'.'}
+    </P>
+  );
+}
+
 function ActionCaption({
   creationError,
   encrypted,
@@ -36,7 +60,7 @@ function ActionCaption({
   creationError: ParsedErrorData | null;
 }) {
   const { textStyles } = useTheme();
-  const { walletName } = useApi();
+  const { walletName, program } = useApi();
 
   if (creationError && creationError.type !== 'DISCONNECTED_FROM_CHAIN') {
     return (
@@ -74,27 +98,113 @@ function ActionCaption({
   return null;
 }
 
+const showAddressInputStatus = (
+  valid: boolean,
+  address: string,
+  cardinalAddress: string
+) => {
+  const { textStyles } = useTheme();
+
+  if (!address) return CardinalCTA();
+
+  const isTwitter = address.charAt(0) === '@';
+
+  if (isTwitter && valid && cardinalAddress) {
+    return (
+      <P
+        className={clsx(textStyles.small, 'dt-text-green-500 dt-mt-1 dt-px-2')}
+      >
+        {cardinalAddress}
+      </P>
+    );
+  }
+
+  if (isTwitter && !valid && !cardinalAddress) {
+    return (
+      <P className={clsx(textStyles.small, 'dt-text-red-500 dt-mt-1 dt-px-2')}>
+        No address is associated with this twitter handle
+      </P>
+    );
+  }
+
+  if (!isTwitter && !valid) {
+    return (
+      <P className={clsx(textStyles.small, 'dt-text-red-500 dt-mt-1 dt-px-2')}>
+        Invalid address or Twitter handle
+      </P>
+    );
+  }
+
+  if (!isTwitter && valid && !cardinalAddress) {
+    return (
+      <P
+        className={clsx(textStyles.small, 'dt-text-green-500 dt-mt-1 dt-px-2')}
+      >
+        Valid address
+      </P>
+    );
+  }
+
+  if (!isTwitter && valid && cardinalAddress) {
+    return (
+      <P
+        className={clsx(textStyles.small, 'dt-text-green-500 dt-mt-1 dt-px-2')}
+      >
+        {cardinalAddress}
+      </P>
+    );
+  }
+};
+
 export default function CreateThread({
   inbox,
   onNewThreadCreated,
   onCloseRequest,
   onModalClose,
 }: CreateThreadProps) {
-  const { createDialect, isDialectCreating, creationError, setDialectAddress } =
-    useDialect();
+  const {
+    createDialect,
+    dialects,
+    isDialectCreating,
+    creationError,
+    setDialectAddress,
+  } = useDialect();
   const { program, network, wallet, walletName } = useApi();
   const { balance } = useBalance();
   const { colors, outlinedInput, textStyles, icons } = useTheme();
 
   const [address, setAddress] = useState('');
+  const [cardinalAddress, setCardinalAddress] = useState('');
   const [encrypted, setEncrypted] = useState(false);
+  const [validAddress, setValidAddress] = useState(false);
+  const [isTyping, setIsTyping] = useState(false);
 
   const createThread = async () => {
-    createDialect(address, [true, true], [false, true], encrypted)
+    const currentChatWithAddress = dialects.find((subscription) => {
+      const otherMemebers = subscription?.dialect.members.filter(
+        (member) =>
+          member.publicKey.toString() !== wallet?.publicKey?.toString()
+      );
+      return (
+        otherMemebers.length == 1 &&
+        address == otherMemebers[0]?.publicKey.toString()
+      );
+    });
+    if (currentChatWithAddress) {
+      const currentThreadWithAddress =
+        currentChatWithAddress.publicKey.toBase58();
+      setDialectAddress(currentThreadWithAddress);
+      onNewThreadCreated?.(currentThreadWithAddress);
+      onCloseRequest?.();
+      return;
+    }
+
+    const finalAddress = cardinalAddress || address;
+    createDialect(finalAddress, [true, true], [false, true], encrypted)
       .then(async () => {
         const [da, _] = await getDialectAddressWithOtherMember(
           program,
-          new anchor.web3.PublicKey(address)
+          new anchor.web3.PublicKey(finalAddress)
         );
         setDialectAddress(da.toBase58());
         onNewThreadCreated?.(da.toBase58());
@@ -104,6 +214,50 @@ export default function CreateThread({
         console.log('error creating dialect', err);
       });
   };
+
+  const onAddressChange = (addr: string) => {
+    setAddress(addr);
+    setIsTyping(true);
+  };
+
+  const tryFetchAddressFromTwitterHandle = async (handle: string) => {
+    const { result } = await fetchAddressFromTwitterHandle(
+      program?.provider.connection,
+      handle
+    );
+
+    if (result) {
+      setValidAddress(true);
+      setCardinalAddress(result.parsed.data.toBase58());
+    } else {
+      setValidAddress(false);
+      setCardinalAddress('');
+    }
+  };
+
+  useEffect(() => {
+    const delayDebounceFn = setTimeout(async () => {
+      if (address.length === 0) {
+        setValidAddress(false);
+      } else if (address.charAt(0) != '@') {
+        try {
+          new anchor.web3.PublicKey(address);
+          setValidAddress(true);
+          setCardinalAddress('');
+        } catch (e) {
+          await tryFetchAddressFromTwitterHandle(address);
+        }
+      } else {
+        const handle = address.substring(1, address.length);
+        await tryFetchAddressFromTwitterHandle(handle);
+      }
+      setIsTyping(false);
+    }, 1000);
+
+    return () => clearTimeout(delayDebounceFn);
+  }, [address]);
+
+  const disabled = !address || (!isTyping && !validAddress);
 
   return (
     <div className="dt-flex dt-flex-col dt-flex-1">
@@ -127,7 +281,7 @@ export default function CreateThread({
         </div>
       </div>
 
-      <div className="dt-h-full dt-pb-8 dt-max-w-sm dt-m-auto dt-flex dt-flex-col">
+      <div className="dt-flex-1 dt-pb-8 dt-max-w-sm dt-m-auto dt-flex dt-flex-col">
         <H1
           className={clsx(
             textStyles.h1,
@@ -138,12 +292,15 @@ export default function CreateThread({
           Create thread
         </H1>
         <Input
-          className={clsx(outlinedInput, 'dt-w-full dt-mb-2')}
-          placeholder="Enter recipient address"
+          className={clsx(outlinedInput, 'dt-w-full dt-mb-1')}
+          placeholder="D1AL...DY5h or @saydialect"
           type="text"
           value={address}
-          onChange={(e) => setAddress(e.target.value)}
+          onChange={(e) => onAddressChange(e.target.value)}
         />
+        {!isTyping
+          ? showAddressInputStatus(validAddress, address, cardinalAddress)
+          : CardinalCTA()}
         <ValueRow
           label={
             <>
@@ -191,11 +348,19 @@ export default function CreateThread({
               />
             </span>
           </ValueRow>
-          <Button onClick={createThread} loading={isDialectCreating}>
+          <Button
+            onClick={createThread}
+            loading={isDialectCreating}
+            disabled={disabled}
+          >
             {isDialectCreating ? 'Creating...' : 'Create thread'}
           </Button>
         </div>
         <ActionCaption encrypted={encrypted} creationError={creationError} />
+      </div>
+
+      <div>
+        <Footer />
       </div>
     </div>
   );