+++ packages/dialect-react-ui/components/ChatButton/index.tsx
@@ -77,7 +77,7 @@ function WrappedChatButton(
   );
   useEffect(() => setRpcUrl(props.rpcUrl || null), [props.rpcUrl, setRpcUrl]);
 
-  const { colors, bellButton, icons, modalWrapper } = useTheme();
+  const { colors, bellButton, icons, modalWrapper, animations } = useTheme();
 
   return (
     <div
@@ -96,16 +96,7 @@ function WrappedChatButton(
         icon={<icons.chat className={cs('dt-w-6 dt-h-6 dt-rounded-full')} />}
         onClick={() => setOpen(!open)}
       ></IconButton>
-      <Transition
-        className={modalWrapper}
-        show={open}
-        enter="dt-transition-opacity dt-duration-300"
-        enterFrom="dt-opacity-0"
-        enterTo="dt-opacity-100"
-        leave="dt-transition-opacity dt-duration-100"
-        leaveFrom="dt-opacity-100"
-        leaveTo="dt-opacity-0"
-      >
+      <Transition className={modalWrapper} show={open} {...animations.popup}>
         <div
           ref={wrapperRef}
           className="dt-w-full dt-h-full"