+++ packages/dialect-react-ui/components/Icon/BackArrow.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 
 const SvgBackArrow = (props: SVGProps<SVGSVGElement>) => (