+++ packages/dialect-react-ui/components/Icon/Chevron.tsx
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import type { SVGProps } from 'react';
 
 const SvgChevron = (props: SVGProps<SVGSVGElement>) => (