+++ packages/dialect-react-ui/components/common/preflighted.tsx
@@ -64,6 +64,13 @@ export const Textarea = ({
   return <textarea className={cs('dt-textarea', className)} {...props} />;
 };
 
+export const Img = ({
+  className,
+  ...props
+}: React.ComponentPropsWithoutRef<'img'> & { className?: string }) => {
+  return <img className={cs('dt-img', className)} {...props} />;
+};
+
 type ButtonProps = React.ComponentPropsWithoutRef<'button'> & {
   className?: string;
 };