+++ packages/dialect-react-ui/components/common/primitives.tsx
@@ -8,10 +8,10 @@ import { ownerFetcher } from '@dialectlabs/web3';
 import { useApi } from '@dialectlabs/react';
 import useSWR from 'swr';
 import cs from '../../utils/classNames';
-import { DialectLogo } from '../Icon';
 import { useTheme } from './ThemeProvider';
 import { A, ButtonBase, P } from './preflighted';
 import clsx from 'clsx';
+import { DialectLogo } from '../Icon';
 
 // TODO: separate these components to separate files
 export function Divider(props: { className?: string }): JSX.Element {
@@ -46,22 +46,23 @@ export function Footer(): JSX.Element {
   const { colors } = useTheme();
 
   return (
-    <div
-      className={cs(
-        'dt-w-[8.5rem] dt-py-1 dt-inline-flex dt-items-center dt-justify-center dt-absolute dt-bottom-3 dt-left-0 dt-right-0 dt-mx-auto dt-uppercase dt-rounded-full',
-        colors.highlightSolid
-      )}
-      style={{ fontSize: '10px' }}
-    >
-      Powered by{' '}
-      <A
-        href="https://dialect.to"
-        target="_blank"
-        rel="noreferrer"
-        className="hover:dt-text-inherit"
+    <div className="dt-flex dt-justify-center dt-py-3">
+      <div
+        className={cs(
+          'dt-px-3 dt-py-1 dt-inline-flex dt-items-center dt-justify-center dt-uppercase dt-rounded-full dt-text-[10px]',
+          colors.highlightSolid
+        )}
       >
-        <DialectLogo className="-dt-mr-1 -dt-mt-1 dt-ml-[3px]" />
-      </A>
+        Powered by{' '}
+        <A
+          href="https://dialect.to"
+          target="_blank"
+          rel="noreferrer"
+          className="hover:dt-text-inherit"
+        >
+          <DialectLogo className="dt-ml-[3px] -dt-mt-[1px]" />
+        </A>
+      </div>
     </div>
   );
 }
@@ -98,13 +99,15 @@ export function Button(props: {
   const { button, buttonLoading, textStyles } = useTheme();
   const defaultClassName = props.defaultStyle || button;
   const loadingClassName = props.loadingStyle || buttonLoading;
+  const disabledClassName =
+    'dt-bg-white dt-text-black dt-border dt-border-white dt-opacity-40';
 
   return (
     <ButtonBase
       className={cs(
         'dt-min-w-120 dt-px-4 dt-py-2 dt-rounded-lg dt-transition-all dt-flex dt-flex-row dt-items-center dt-justify-center',
         textStyles.buttonText,
-        defaultClassName,
+        props.disabled ? disabledClassName : defaultClassName,
         props.loading && loadingClassName,
         props.className
       )}
@@ -172,7 +175,7 @@ export function Toggle({
     >
       <input
         type="checkbox"
-        className="dt-appearance-none dt-opacity-0 dt-w-0 dt-h-0"
+        className="dt-input dt-appearance-none dt-opacity-0 dt-w-0 dt-h-0"
         checked={checked}
         onChange={() => {
           setChecked((prev) => !prev);
@@ -215,7 +218,7 @@ export function Accordion(props: {
           textStyles.bigText,
           'dt-w-full dt-flex dt-justify-between dt-mb-1'
         )}
-      > 
+      >
         {props.title}
         {/* <IconButton
           icon={<icons.chevron />}
@@ -231,7 +234,6 @@ export function Accordion(props: {
     </div>
   );
 }
-
 
 export const useBalance = () => {
   const { wallet, program } = useApi();